Add files via upload
First language is the one I want and the second one my input texts language.
Why this order? Because the Al Bhed 'language' varies between Spanish, English, Japanese and others.

That way albhed_'lang'.py is better for organization when translating from 'lang' to Al Bhed. The reverse order is the reverse action.
Japanese and Spanish are next!